check cflags

+++ src_yf/main/env_utils.c
@@ -64,8 +64,8 @@ char	*get_env_content(char *full, char *var_name)
 	int		i;
 	int		j;
 	char	*content;
-	int		content_len;
-	int		var_name_len;
+	size_t	content_len;
+	size_t	var_name_len;
 
 	var_name_len = ft_strlen(var_name);
 	if (var_name_len + 1 == ft_strlen(full))
